Clear localStorage between tests
Not doing this means that subsequent tests are affected by the state of
earlier tests. Although we could also inject the `localStorage`
dependency into the `TodoApp` component, this would mean that that the
browser tests running on Karma wouldn't validate the various browser
implementations of `localStorage` that we want to test against.