Make the router an injected property
So that we can test changes to the page URL within tests, we will
instead need to dispatch events to the router, which of course requires
access to the router from within the test. I also added class names to
hyperlinks so they can be accessed without using brittle selectors.